requirements.txt
APScheduler == 3.7.0
bcrypt == 3.2.0
boto3 == 1.17.5
botocore == 1.20.5
cryptography == 3.4.6
docopt == 0.6.2
Flask == 1.1.2
Flask-Login == 0.5.0
graphene == 3.0b7
graphql-core == 3.1.2
graphql-relay == 3.0.0
graphql-server-core == 2.0.0
Jinja2 == 2.11.3
jsonmerge == 1.7.0
jsonschema == 3.2.0
lxml == 4.6.2
lz4 == 2.2.1
networkx == 2.5
ordered-set == 4.0.2
paramiko == 2.7.2
pendulum == 2.1.2
pony == 0.7.14
protobuf == 3.12.4
psutil == 5.8.0
pycrypto == 2.6.1
python-dateutil == 2.8.1
python-snappy == 0.6.0
pytz == 2021.1
PyYAML == 5.3.1
requests == 2.25.1
setuptools == 53.1.0
urllib3 == 1.26.3
zstandard == 0.15.2
